receipt-thesis-backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pandas
scikit-learn
joblib
pytesseract
opencv-python-headless
Pillow
streamlit
tqdm
python-Levenshtein
python-dateutil
ultralytics
SQLAlchemy
python-dotenv
httpx
psycopg2-binary

